requirements.txt
langchain
qdrant-client
psycopg2-binary
pdfplumber
transformers
sentence-transformers
qwen-embedding
fastapi
uvicorn
python-dotenv
